Lift the account state to App and drop redundant comments
Review feedback: the card reporting its derived address upward through
an onAddressChange effect was the wrong abstraction. The account is
app-level state — the header shows it, the card trades with it — so App
now owns it and passes it down; the card mutates the state it was given
through onAccountChange. The AccountState union and an accountAddress
helper move to account.ts beside the cache it hydrates from. Also drops
the NETWORK_NAME rationale comment and the AccountChip prop JSDoc.
